+++ CHANGELOG.md
@@ -5,6 +5,28 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.2.2] - 2023-03-30
+
+### Added
+
+- Action version logging
+
+### Changed
+
+- Update all external NPM dependencies
+
+## [2.2.1] - 2023-03-29
+
+### Fixed
+
+- Error: The auto scaling configuration ARN you specified isn't valid. ARN: '' ([#36](https://github.com/awslabs/amazon-app-runner-deploy/issues/36))
+
+## [2.2.0] - 2023-03-28
+
+### Added
+
+- Optional `auto-scaling-config-arn` input. When specified this will be applied to new and existing services.
+
 ## [2.1.0] - 2023-02-08
 
 ### Added